+++ miden-crypto/src/merkle/smt/large_forest/property_tests.rs
@@ -51,7 +51,14 @@ proptest! {
         // Iterating over the historical version of the lineage in the forest should produce exactly
         // the same entries as iterating over V1 of our test tree.
         let old_version = TreeId::new(lineage, version);
-        let forest_entries = forest.entries(old_version).map_err(to_fail)?.sorted().collect_vec();
+        let forest_entries = forest
+            .entries(old_version)
+            .map_err(to_fail)?
+            .collect::<Result<Vec<_>, _>>()
+            .map_err(to_fail)?
+            .into_iter()
+            .sorted()
+            .collect_vec();
         let tree_entries = tree_v1
             .entries()
             .map(|(k, v)| TreeEntry { key: *k, value: *v })
@@ -66,7 +73,14 @@ proptest! {
         } else {
             TreeId::new(lineage, tree_info.version())
         };
-        let forest_entries = forest.entries(current_version).map_err(to_fail)?.sorted().collect_vec();
+        let forest_entries = forest
+            .entries(current_version)
+            .map_err(to_fail)?
+            .collect::<Result<Vec<_>, _>>()
+            .map_err(to_fail)?
+            .into_iter()
+            .sorted()
+            .collect_vec();
         let tree_entries = tree_v2
             .entries()
             .map(|(k, v)| TreeEntry { key: *k, value: *v })
@@ -92,7 +106,12 @@ proptest! {
         // Iterating over the historical version of the lineage in the forest should produce exactly
         // the same entries as iterating over V1 of our test tree.
         let old_version = TreeId::new(lineage, version);
-        prop_assert!(forest.entries(old_version).map_err(to_fail)?.all(|e| e.value != EMPTY_WORD));
+        let entries = forest
+            .entries(old_version)
+            .map_err(to_fail)?
+            .collect::<Result<Vec<_>, _>>()
+            .map_err(to_fail)?;
+        prop_assert!(entries.iter().all(|e| e.value != EMPTY_WORD), "EMPTY_WORD entry encountered");
 
         // Iterating over the newest version of the lineage in the forest should provide exactly the
         // same entries as iterating over V2 of our test tree.
@@ -101,6 +120,11 @@ proptest! {
         } else {
             TreeId::new(lineage, root_2.version())
         };
-        prop_assert!(forest.entries(current_version).map_err(to_fail)?.all(|e| e.value != EMPTY_WORD));
+        let entries = forest
+            .entries(current_version)
+            .map_err(to_fail)?
+            .collect::<Result<Vec<_>, _>>()
+            .map_err(to_fail)?;
+        prop_assert!(entries.iter().all(|e| e.value != EMPTY_WORD), "EMPTY_WORD entry encountered");
     }
 }